Rename package to lrsctrl
