+++ lrsctrl/cli.py
@@ -1,29 +1,29 @@
 import click
-from adc64.client import Client
-from adc64 import server
+from lrsctrl.client import Client
+from lrsctrl import server
 
 
 @click.group()
-def adc64():
+def lrsctrl():
     pass
 
 
-@adc64.command()
+@lrsctrl.command()
 def start():
     Client().send_start_adc64()
 
-@adc64.command()
+@lrsctrl.command()
 def stop():
     Client().send_stop_adc64()
 
-@adc64.command()
+@lrsctrl.command()
 def start_rc():
     Client().send_start_rc()
 
-@adc64.command()
+@lrsctrl.command()
 def stop_rc():
     Client().send_stop_rc()
 
-@adc64.command()
+@lrsctrl.command()
 def serve():
     server.start_app()
