+++ lrsctrl/client.py
@@ -1,5 +1,5 @@
 import requests
-from adc64.config import Config
+from lrsctrl.config import Config
 
 
 class Client():
@@ -14,25 +14,25 @@ class Client():
         try:
             requests.get(addr)
         except:
-            print(f'Server disconected! Use CLI command to CLI-server: adc64 serve')
+            print(f'Server disconected! Use CLI command to CLI-server: lrsctrl serve')
 
     def send_stop_adc64(self):
         addr = f'{self.url}/api/stop_adc64'
         try:
             requests.get(addr)
         except:
-            print(f'Server disconected! Use CLI command to CLI-server: adc64 serve')
+            print(f'Server disconected! Use CLI command to CLI-server: lrsctrl serve')
 
     def send_start_rc(self):
         addr = f'{self.url}/api/start_rc'
         try:
             requests.get(addr)
         except:
-            print(f'Server disconected! Use CLI command to CLI-server: adc64 serve')
+            print(f'Server disconected! Use CLI command to CLI-server: lrsctrl serve')
 
     def send_stop_rc(self):
         addr = f'{self.url}/api/stop_rc'
         try:
             requests.get(addr)
         except:
-            print(f'Server disconected! Use CLI command to CLI-server: adc64 serve')
+            print(f'Server disconected! Use CLI command to CLI-server: lrsctrl serve')
